requirements.txt
discord
jsonschema
numpy
pillow
requests
scons
